backend/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.10
elevenlabs>=0.2.0
sentence-transformers>=2.2.0
groq>=0.4.0
python-dotenv>=0.19.0
playsound>=1.2.2
flask>=2.0.0
flask-cors>=3.0.0
SpeechRecognition>=3.10.0